requirements.txt
# ============================================================
# Requirements for Image Forgery Detection Project
# Deep Learning Based Image Forgery Detection Using CNNs
# ============================================================

# Image Processing
Pillow>=9.5.0
opencv-python>=4.7.0

# Scientific Computing
numpy>=1.24.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.22.0

# User Interface
streamlit>=1.24.0
